fix: 修复了解析json时数据类型错误的bug

+++ src/main/kotlin/cn/rtast/rmusic/models/netease/detail/ChargeInfo.kt
@@ -8,7 +8,7 @@ package cn.rtast.rmusic.models.netease.detail
 
 data class ChargeInfo(
     val chargeMessage: Any,
-    val chargeType: Int,
+    val chargeType: Long,
     val chargeUrl: Any,
-    val rate: Int
+    val rate: Long
 )
