+++ src/main/kotlin/cn/rtast/rmusic/models/netease/search/Result.kt
@@ -8,6 +8,6 @@ package cn.rtast.rmusic.models.netease.search
 
 data class Result(
     val hasMore: Boolean,
-    val songCount: Int,
+    val songCount: Long,
     val songs: List<Song>
 )
